// pos/src/app/models/m-warehouse-master.ts
export class MWarehouseMaster {
constructor(){}

    id?: number;
    warehouseCode?: string;
    warehouseName?: string;
    countryID?: number;
    companyID?: number;
    warehousetypeID?: number;
    countryName?: string;
    companyName?: string;
    warehouseTypeName?: string;
    rowNumber?: string;
    remarks?: string;
    active?: boolean;
    countryCode?: string;
    companyCode?: string;
    warehousetypeCode?: string;
}